+ сервис для вычитсления оплаты. Добавлено поле "ед измерения" в таблицу Услуги. Необходимо пересоздать таблицы.

+++ src/main/kotlin/ru/kotlinschool/service/UserService.kt
@@ -19,7 +19,7 @@ interface UserService {
     /**
      * Добавление квартиры
      */
-    fun registerFlat(userId: Long, houseId: Long, flatNumber: String, area: Double, numberOfResidents: Int)
+    fun registerFlat(userId: Long, houseId: Long, flatNumber: String, area: Double, numberOfResidents: Long)
 
     /**
      * Получить все квартиры пользователя
